% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{bbox_handler}
\alias{bbox_handler}
\title{Mikey lurnin}
\usage{
bbox_handler()
}
\description{
https://paleolimbot.github.io/wk/articles/programming.html
}
\examples{
bbox_handler()
library(wk)
wk_handle(
  wkt(c(NA, "POINT (0 1)", "GEOMETRYCOLLECTION (POINT (2 3))")),
  bbox_handler()
)

# The most important thing about a wk_handler is that, once created, it can only be used once:
h <- bbox_handler()
wk_handle(wkt(), h)
try(wk_handle(wkt(), h))

## pass a function for user-supplied handles
handler <- function() bbox_handler()
wk_handle(wkt(), handler)
wk_handle(wkt(), handler)

}
