% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk.programming-package.R
\docType{package}
\name{wk.programming-package}
\alias{wk.programming}
\alias{wk.programming-package}
\title{wk.programming: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
